Problem: a potential lifetime problem with Block and Blockchain Solution: move the blockhash function to Block
It is possible for Block and Blockchain to contain different lifetime
in the current implementation. The correct way to handle this is to
either a) use two different lifetimes or b) merge them into the same
trait.

Solution a) will unnecessarily complicate things, as there will
possibly ever be only one function for Blockchain -- the blockhash
function. So b) might be a better solution.