Problem: create a cache in machine VM creates potential problem Solution: delay implementation of storage to block
Think of a chain of machine forks. The top level machine executes on
account A, which then triggers another machine fork that executes on
account B, which then triggers a third machine fork that executes on
account A. There's no way -- if we implement the storage cache in
Machine -- to track how to update the storage cache when those two
forks finish execution. Those information should be on the current
block.